try 2

+++ 02/parser.hpp
@@ -8,7 +8,7 @@ using namespace std;
 class TokenParser
 {
 private:
-    function<string (int)> DigitToken = nullptr;
+    function<string (uint64_t)> DigitToken = nullptr;
     function<string (const string&)> StringToken = nullptr;
     function<string ()> StartToken = nullptr;
     function<string ()> EndToken = nullptr;
@@ -16,7 +16,7 @@ private:
 public:
     void SetStartCallback(const function<string ()>& start);
     void SetEndCallback(const function<string ()>& end);
-    void SetDigitTokenCallback(const function<string (int)>& digit);
+    void SetDigitTokenCallback(const function<string (uint64_t)>& digit);
     void SetStringTokenCallback(const function<string (const string&)>& string);
     string Parser(const string & text);
     string ConvertToken(const string & text);
